refactor

+++ backend/src/main/java/com/example/training/entity/Account.java
@@ -9,6 +9,7 @@ import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.Range;
 
 import javax.persistence.*;
+import javax.validation.constraints.Digits;
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.Size;
@@ -33,10 +34,10 @@ public class Account {
     private Long accNo;
 
     @Column(name="userId")
-    private Integer userId;
+    private Long userId;
 
     @Column(name="aadhaarNo")
-    private long aadhaarNo;
+    private Long aadhaarNo;
 
     @Column(name="branch")
     private String branch;
